Move SizedUnsignedTrait to StdLibExtras.h and remove dupes https://bugs.webkit.org/show_bug.cgi?id=300822 rdar://162706445
Reviewed by Keith Miller.

We have three copies of this. Move to a generic header.

* Source/JavaScriptCore/parser/Lexer.cpp:
(JSC::Lexer<T>::parseString):
(JSC::Lexer<T>::lexWithoutClearingLineTerminator):
* Source/JavaScriptCore/runtime/JSONObject.cpp:
(JSC::stringCopySameType):
* Source/JavaScriptCore/runtime/LiteralParser.cpp:
(JSC::reviverMode>::Lexer::lexString):
* Source/WTF/wtf/EnumSet.h:
* Source/WTF/wtf/HashFunctions.h:
(WTF::IntHash::hash):
* Source/WTF/wtf/SIMDHelpers.h:
* Source/WTF/wtf/SortedArrayMap.h:
(WTF::foldForComparison):
* Source/WTF/wtf/StdLibExtras.h:
* Source/WTF/wtf/text/StringCommon.h:
(WTF::charactersContain):
(WTF::countMatchedCharacters):

Canonical link: https://commits.webkit.org/301581@main
